feat: Support multiple categories for forms
Updates the `FormDefinition` type to accept an array of `Category` objects, allowing forms to be associated with multiple categories instead of just one. This change also includes migration logic to convert existing single-category forms to the new array format. Additionally, a new modal to display the list of available forms has been introduced.